% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{fallback}
\alias{fallback}
\title{Create a fallback}
\usage{
fallback(terminal_fallback_value, source_file = "config.yaml",
  hierarchy = c("./", "~/"), key = NULL)
}
\arguments{
\item{terminal_fallback_value}{The value the fallback will resolve to when
no other instance in the hierarchy defines the key.}

\item{hierarchy, source_file}{Define the paths to the yaml files where we
look for a key to be defined. By default, it is a file called \code{config.yaml}
in the working directory and the home directory.}

\item{key}{The key, that is, the name of the argument the fallback is created
for. Can be \code{NULL}, so the definition of the key is deferred until
\code{\link[=resolve_fallback]{resolve_fallback()}} is called to avoid redundancy.}
}
\description{
Creates a fallback that can be evaluated with \code{\link[=resolve_fallback]{resolve_fallback()}}.
}
\examples{
fallback(TRUE)
}
