% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve_fallback}
\alias{resolve_fallback}
\title{Resolve a fallback}
\usage{
resolve_fallback(fallback)
}
\arguments{
\item{fallback}{A fallback chain defined with \code{\link[=fallback]{fallback()}}.}
}
\description{
Resolves the value of a fallback chain defined with \code{\link[=fallback]{fallback()}}.
}
\examples{

f <- function(x = fallback(TRUE)) {
  resolve_fallback(x)$value
}

f() # with no config files in place, this resolves to the terminal fallback.
dir <- tempdir()
dir1 <- fs::path(dir, "dir1")
fs::dir_create(dir1)
yaml::write_yaml(list(frog = 100), fs::path(dir1, "config.yaml"))


g <- function(frog = fallback(letters, hierarchy = dir1)) {
  resolve_fallback(frog)$value
}

# this should resolve to the fallback declared in dir1
g()
}
